+++ CSharpPICollision/MainWindow.xaml.cs
@@ -33,7 +33,7 @@ namespace CSharpPICollision
             controller.Camera.Transform = controller.Transform;
         }
 
-        private (double Mass, double Speed)? ShowPropertiesEditDialog()
+        private (decimal Mass, decimal Speed)? ShowPropertiesEditDialog()
         {
             var dialog = new PropertiesEditDialog();
             var result = dialog.ShowDialog();
@@ -93,9 +93,9 @@ namespace CSharpPICollision
             return timer;
         }
 
-        private (Block, Block) InitializeBlocks(double mass, double speed)
+        private (Block, Block) InitializeBlocks(decimal mass, decimal speed)
         {
-            return (new Block(1, 1, 0, 2), new Block(1.5, mass, speed, 6));
+            return (new Block(1, 1, 0, 2), new Block(1.5m, mass, speed, 6));
         }
 
         private (IPhysicalEngine physicalEngine, IVisualEngine visualEngine) InitializeEngines(CameraController cameraController, Viewport3D viewport, (Block, Block) blocks)
@@ -120,7 +120,7 @@ namespace CSharpPICollision
             dispatcherTimer.Stop();
         }
 
-        public MainWindow(double mass, double speed)
+        public MainWindow(decimal mass, decimal speed)
         {
             InitializeComponent();
 
@@ -141,7 +141,7 @@ namespace CSharpPICollision
             Start(engines.physicalEngine, dispatcherTimer);
         }
 
-        public MainWindow() : this(1e6, -2)
+        public MainWindow() : this(1e6m, -2)
         {
 
         }
@@ -170,7 +170,7 @@ namespace CSharpPICollision
 
         private void Window_MouseWheel(object sender, System.Windows.Input.MouseWheelEventArgs e)
         {
-            cameraController.Scale(-0.1 * Math.Sign(e.Delta) + 1);
+            cameraController.Scale(-0.1m * Math.Sign(e.Delta) + 1);
 
             SetCameraProperties(cameraController);
         }
